Emmiting ON_FINISH_LOADING_MAP event when map finishes loading AND map is ready
Solves the issue #586. When onMapFinishedLoading is called user would request LatLng bounds,
but it is not always possible at that points since onMapReady in java code is still not called,
and _map is null. Previouse in that case user would get ugly red screen with exception displayed,
and could do nothing about it. Now user can be sure that when onMapFinishedLoading is called,
map is completely ready.

I don't like introducing new flags, but could not think about better solution.
At least two flags introduced in my commits are self describing, and their purpose is clear.